[xabt] Rename PostTrimmingRemoveResourceDesignerStep to RemoveResourceDesignerStep
The old ILLink RemoveResourceDesignerStep is no longer compiled into
Xamarin.Android.Build.Tasks, so the PostTrimming prefix is unnecessary.
Also remove AndroidLinkConfiguration.cs from compilation as it was only
used by the old ILLink step.
