+++ src/Xamarin.Android.Build.Tasks/Xamarin.Android.Build.Tasks.csproj
@@ -52,13 +52,11 @@
     <Compile Include="Linker\MonoDroid.Tuner\AddKeepAlivesHelper.cs" />
     <Compile Include="Linker\MonoDroid.Tuner\StripEmbeddedLibrariesStep.cs" />
     <Compile Include="Linker\MonoDroid.Tuner\PostTrimmingAddKeepAlivesStep.cs" />
-    <Compile Include="Linker\MonoDroid.Tuner\PostTrimmingRemoveResourceDesignerStep.cs" />
-    <Compile Include="Linker\MonoDroid.Tuner\AndroidLinkConfiguration.cs" />
+    <Compile Include="Linker\MonoDroid.Tuner\RemoveResourceDesignerStep.cs" />
     <Compile Include="Linker\MonoDroid.Tuner\Extensions.cs" />
     <Compile Include="Linker\MonoDroid.Tuner\FixAbstractMethodsStep.cs" />
     <Compile Include="Linker\MonoDroid.Tuner\FixLegacyResourceDesignerStep.cs" />
     <Compile Include="Linker\MonoDroid.Tuner\LinkDesignerBase.cs" />
-    <Compile Include="Linker\MonoDroid.Tuner\RemoveResourceDesignerStep.cs" />
     <Compile Include="Linker\External\Linker\Annotations.cs" />
     <Compile Include="Linker\External\Linker\AssemblyAction.cs" />
     <Compile Include="Linker\External\Linker\AssemblyResolver.cs" />
